requirements.txt
jupyter>=1.0.0
numpy
openai-clip
pandas>=1.3.4
Pillow
requests>=2.22.0
torch
torchvision
tqdm
transformers